% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-slack.R
\name{.has_reaction}
\alias{.has_reaction}
\title{Check for a Reaction in Reactions}
\usage{
.has_reaction(reactions, reaction_name, posters, valid_users)
}
\arguments{
\item{reactions}{A list of reaction lists.}

\item{reaction_name}{A character vector with the name of one or more target
reactions.}

\item{posters}{An optional character vector of user ids of the person who
posted each message. If provided, reactions from this user count as valid.}

\item{valid_users}{An optional character vector of user ids whose reactions
always count as valid. If this and posters are both omitted, all reactions
are considered valid.}
}
\value{
A logical vector the same length as reactions.
}
\description{
Check for a Reaction in Reactions
}
\keyword{internal}
