% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4ds-tags.R
\name{r4ds_process_convos}
\alias{r4ds_process_convos}
\title{Add R4DS-specific information}
\usage{
r4ds_process_convos(convos_tbl)
}
\arguments{
\item{convos_tbl}{Tidy conversations.}
}
\value{
A tbl with additional column \code{status}. Probably other things soon.
}
\description{
For R4DS, we use an emoji language to tag questions, and combine that
information with other post properties to assign a status.
}
