% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4dsposts-package.R
\docType{package}
\name{r4dsposts-package}
\alias{r4dsposts}
\alias{r4dsposts-package}
\title{r4dsposts: Archive and Analyze Posts on R4DS}
\description{
r4dsposts will be used to create an archive of posts on the R4DS Slack.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r4ds/r4dsposts}
  \item \url{https://r4ds.github.io/r4dsposts/}
  \item Report bugs at \url{https://github.com/r4ds/r4dsposts/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
